requirements_ng.txt
Pillow==3.0.0
pygame==1.9.2a0
